requirements.txt
beautifulsoup4==4.9.3
jieba==0.42.1
Jinja2==2.11.2
markupsafe==2.0.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.2
pandas==1.1.4
pynvml==8.0.4
scikit-learn==0.23.2
scipy==1.5.4
spacy==2.3.4
tensorboard==2.4.0
TorchSnooper==0.8
tqdm==4.56.0
transformers==4.6.1
protobuf==3.19.3